refactor(skillet): implement host-specific configuration and pihole crate
