+++ ublue/skillet/crates/cli-common/src/lib.rs
@@ -1,8 +1,7 @@
 use clap::Parser;
-use skillet_core::files::LocalFileResource;
+use skillet_core::files::{FileResource, LocalFileResource};
 use skillet_core::recorder::Recorder;
-use skillet_core::system::LinuxSystemResource;
-use skillet_hardening::apply;
+use skillet_core::system::{LinuxSystemResource, SystemResource};
 use std::fs;
 use std::path::PathBuf;
 use thiserror::Error;
@@ -11,8 +10,8 @@ use tracing_subscriber::FmtSubscriber;
 
 #[derive(Error, Debug)]
 pub enum CliCommonError {
-    #[error("Failed to apply hardening: {0}")]
-    Hardening(#[from] skillet_hardening::HardeningError),
+    #[error("Configuration error: {0}")]
+    Config(String),
     #[error("System error: {0}")]
     System(#[from] skillet_core::system::SystemError),
     #[error("Failed to set default tracing subscriber: {0}")]
@@ -44,7 +43,10 @@ pub enum HostCommands {
     },
 }
 
-pub fn run_host(hostname: &str) -> Result<(), CliCommonError> {
+pub fn run_host<F>(hostname: &str, apply_fn: F) -> Result<(), CliCommonError>
+where
+    F: Fn(&dyn SystemResource, &dyn FileResource) -> Result<(), String>,
+{
     let args = HostArgs::parse();
 
     let subscriber = FmtSubscriber::builder()
@@ -58,11 +60,18 @@ pub fn run_host(hostname: &str) -> Result<(), CliCommonError> {
     tracing::subscriber::set_global_default(subscriber)?;
 
     match args.command {
-        HostCommands::Apply { record } => handle_apply(hostname, record),
+        HostCommands::Apply { record } => handle_apply(hostname, record, apply_fn),
     }
 }
 
-pub fn handle_apply(hostname: &str, record_path: Option<PathBuf>) -> Result<(), CliCommonError> {
+pub fn handle_apply<F>(
+    hostname: &str,
+    record_path: Option<PathBuf>,
+    apply_fn: F,
+) -> Result<(), CliCommonError>
+where
+    F: Fn(&dyn SystemResource, &dyn FileResource) -> Result<(), String>,
+{
     info!("Starting Skillet configuration for {}...", hostname);
 
     let system = LinuxSystemResource::new();
@@ -72,20 +81,17 @@ pub fn handle_apply(hostname: &str, record_path: Option<PathBuf>) -> Result<(), 
         let recorder_system = Recorder::new(system);
         let recorder_files = Recorder::with_ops(files, recorder_system.shared_ops());
 
-        apply(&recorder_system, &recorder_files)?;
+        apply_fn(&recorder_system, &recorder_files).map_err(CliCommonError::Config)?;
 
         let ops = recorder_system.get_ops();
         let yaml = serde_yml::to_string(&ops)?;
         if let Some(parent) = path.parent() {
             fs::create_dir_all(parent)?;
         }
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)?;
-        }
         fs::write(&path, yaml)?;
         info!("Recording saved to {}", path.display());
     } else {
-        apply(&system, &files)?;
+        apply_fn(&system, &files).map_err(CliCommonError::Config)?;
     }
 
     info!("Configuration applied successfully.");
