+++ ublue/skillet/crates/hosts/beezelbot/src/main.rs
@@ -2,6 +2,8 @@ use anyhow::Result;
 use skillet_cli_common::run_host;
 
 fn main() -> Result<()> {
-    run_host("beezelbot")?;
+    run_host("beezelbot", |system, files| {
+        skillet_hardening::apply(system, files).map_err(|e| e.to_string())
+    })?;
     Ok(())
 }
