+++ ublue/skillet/crates/hosts/clamps/Cargo.toml
@@ -8,4 +8,6 @@ workspace = true
 
 [dependencies]
 skillet_cli_common.workspace = true
+skillet_hardening.workspace = true
+skillet_pihole.workspace = true
 anyhow = "1.0"
