+++ ublue/skillet/crates/hosts/clamps/src/main.rs
@@ -2,6 +2,10 @@ use anyhow::Result;
 use skillet_cli_common::run_host;
 
 fn main() -> Result<()> {
-    run_host("clamps")?;
+    run_host("clamps", |system, files| {
+        skillet_hardening::apply(system, files).map_err(|e| e.to_string())?;
+        skillet_pihole::apply(system, files).map_err(|e| e.to_string())?;
+        Ok(())
+    })?;
     Ok(())
 }
